Start splitting prsc parser into multiple files
